+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Added
+- Support for Docker private registry - see [#10](https://github.com/ExpediaInc/apiary-authorization/issues/10).
+
 ## [2.0.1] - 2019-03-12
 
 ### Added